Fix rounded rectangle card changing orientation based on number
The appearance of rounded rectangles were not consistent. One rounded rectangle would stretch vertically, two rounded rectangles would appear to be two circles, and three rounded rectangles would stretch horizontally. I believe that adding the .frame() view modifier to the is acceptable in this case because the view is for this specific type of game with rounded rectangle shapes. For games with more shapes, slightly altered types of views would likely have to be constructed.