Implement function that checks if a set is available to be matched.
The function was previously calling isSet() which was a mutating function that I put in charge of dealing out three more cards when a set was made. This was a great lesson on being careful with mutating functions and realizing that they can lead to unexpected results, even if all of your new code seems correct.